Gemfile.lock: Rebuild.
When you conduct a 'bundle install', bundler uses the versions of gems described
in the Gemfile.lock file. However, some of the versions specified in the lockfile
become unavailable, as sometimes developers remove specific versions from
circulation due to security vulnerabilities. Removing them from circulation
removes the chance that someone installs an insecure version. But, folks who
have lockfiles that refer to those verisons must either edit the lockfile
or rebuild the lockfile, which is a simpler approach.

Rebuild the lockfile, so that it contains updated/available versions of all
library dependencies.